# swagger copy.json
{
  "info": {
    "description": "API documentation for ocemQ",
    "version": "1.0.0",
    "title": "OCEM Quiz",
    "contact": {
      "email": "linkwithdm@gmail.com"
    }
  },
  "basePath": "/",
  "swagger": "2.0",
  "tags": [
    {
      "name": "Authentication",
      "description": "Everything about authentication"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Register",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the server",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be compare to the existing server user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["User"],
        "description": "Get All Registered User From Server",
        "summary": "Get All Registered Users",
        "parameters": [],
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/question": {
      "get": {
        "tags": ["Question"],
        "description": "Get All Questions from servers",
        "summary": "Get All Questions",
        "parameters": [],
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/question/checkAnswer/{qid}/{answer}/{username}": {
      "get": {
        "tags": ["Question"],
        "summary": "Check Answer",
        "parameters": [
          {
            "in": "path",
            "name": "qid",
            "type": "number",

            "required": true
          },
          {
            "in": "path",
            "name": "answer",
            "type": "string",

            "required": true
          },
          {
            "in": "path",
            "name": "username",
            "type": "string",

            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scoreboard": {
      "get": {
        "tags": ["Scoreboard"],
        "description": "Get Top Scorer From Server",
        "summary": "Get Scoreboard",
        "parameters": [],
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Register": {
      "type": "object",
      "required": ["name", "username", "password"],
      "properties": {
        "name": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Login": {
      "type": "object",
      "required": ["username", "password"],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Ans": {
      "type": "object",
      "required": ["userId", "answerSelectedByUser"],
      "properties": {
        "userId": {
          "type": "string"
        },
        "answerSelectedByUser": {
          "type": "string"
        }
      }
    }
  }
}